Adds static toll rates to database
Adds repo and view model for the new rates
