--- app/src/main/java/gov/wa/wsdot/android/wsdot/database/tollrates/constant/TollRate.java
@@ -1,9 +1,0 @@
-package gov.wa.wsdot.android.wsdot.database.tollrates.constant;
-
-public class TollRate {
-
-
-
-
-
-}
