+++ app/src/main/java/gov/wa/wsdot/android/wsdot/database/tollrates/dynamic/tollratesign/TollRateSignEntity.java
@@ -1,4 +1,4 @@
-package gov.wa.wsdot.android.wsdot.database.tollrates.dynamic;
+package gov.wa.wsdot.android.wsdot.database.tollrates.dynamic.tollratesign;
 
 import androidx.annotation.NonNull;
 import androidx.room.ColumnInfo;
