+++ app/src/main/java/gov/wa/wsdot/android/wsdot/database/tollrates/dynamic/tollratesign/tolltrips/TollTripDao.java
@@ -1,4 +1,4 @@
-package gov.wa.wsdot.android.wsdot.database.tollrates.dynamic;
+package gov.wa.wsdot.android.wsdot.database.tollrates.dynamic.tollratesign.tolltrips;
 
 import androidx.room.Dao;
 import androidx.room.Insert;
@@ -10,13 +10,13 @@ import androidx.room.Transaction;
 public abstract class TollTripDao {
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    public abstract void insertTravelTimes(TollRateGroupDao.TollTripEntity... tollTrips);
+    public abstract void insertTravelTimes(TollTripEntity... tollTrips);
 
     @Query("DELETE FROM toll_trip")
     public abstract void deleteAll();
 
     @Transaction
-    public void deleteAndInsertTransaction(TollRateGroupDao.TollTripEntity... tollTrips) {
+    public void deleteAndInsertTransaction(TollTripEntity... tollTrips) {
         deleteAll();
         insertTravelTimes(tollTrips);
     }
