+++ app/src/main/java/gov/wa/wsdot/android/wsdot/di/AppModule.java
@@ -17,8 +17,8 @@ import gov.wa.wsdot.android.wsdot.database.mountainpasses.MountainPassDao;
 import gov.wa.wsdot.android.wsdot.database.myroute.MyRouteDao;
 import gov.wa.wsdot.android.wsdot.database.notifications.NotificationTopicDao;
 import gov.wa.wsdot.android.wsdot.database.tollrates.dynamic.TollRateGroupDao;
-import gov.wa.wsdot.android.wsdot.database.tollrates.dynamic.TollRateSignDao;
-import gov.wa.wsdot.android.wsdot.database.tollrates.dynamic.TollTripDao;
+import gov.wa.wsdot.android.wsdot.database.tollrates.dynamic.tollratesign.TollRateSignDao;
+import gov.wa.wsdot.android.wsdot.database.tollrates.dynamic.tollratesign.tolltrips.TollTripDao;
 import gov.wa.wsdot.android.wsdot.database.trafficmap.MapLocationDao;
 import gov.wa.wsdot.android.wsdot.database.traveltimes.TravelTimeDao;
 import gov.wa.wsdot.android.wsdot.database.traveltimes.TravelTimeGroupDao;
