+++ app/src/main/java/gov/wa/wsdot/android/wsdot/di/ViewModelModule.java
@@ -25,7 +25,7 @@ import gov.wa.wsdot.android.wsdot.ui.myroute.report.alerts.MyRouteAlertListViewM
 import gov.wa.wsdot.android.wsdot.ui.myroute.report.cameras.MyRouteCamerasViewModel;
 import gov.wa.wsdot.android.wsdot.ui.myroute.report.traveltimes.MyRouteTravelTimesViewModel;
 import gov.wa.wsdot.android.wsdot.ui.notifications.NotificationsViewModel;
-import gov.wa.wsdot.android.wsdot.ui.tollrates.TollRatesViewModel;
+import gov.wa.wsdot.android.wsdot.ui.tollrates.TollRateSignsViewModel;
 import gov.wa.wsdot.android.wsdot.ui.trafficmap.FavoriteMapLocationViewModel;
 import gov.wa.wsdot.android.wsdot.ui.trafficmap.alertsinarea.HighwayAlertListViewModel;
 import gov.wa.wsdot.android.wsdot.ui.trafficmap.alertsinarea.HighwayAlertViewModel;
@@ -196,8 +196,8 @@ abstract class ViewModelModule {
 
     @Binds
     @IntoMap
-    @ViewModelKey(TollRatesViewModel.class)
-    abstract ViewModel bindI405TollRatesViewModel(TollRatesViewModel i405TollRatesViewModel);
+    @ViewModelKey(TollRateSignsViewModel.class)
+    abstract ViewModel bindI405TollRatesViewModel(TollRateSignsViewModel i405TollRatesViewModel);
 
     @Binds
     abstract ViewModelProvider.Factory bindViewModelFactory(ViewModelFactory factory);
