+++ app/src/main/java/gov/wa/wsdot/android/wsdot/repository/TollRatesRepository.java
@@ -1,33 +1,31 @@
 package gov.wa.wsdot.android.wsdot.repository;
 
 import android.text.format.DateUtils;
-import android.util.Log;
+
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
 
 import org.json.JSONArray;
-import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
 import java.net.URL;
 import java.net.URLConnection;
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
+import gov.wa.wsdot.android.wsdot.database.Converters;
 import gov.wa.wsdot.android.wsdot.database.caches.CacheEntity;
-import gov.wa.wsdot.android.wsdot.database.tollrates.dynamic.TollRateGroup;
-import gov.wa.wsdot.android.wsdot.database.tollrates.dynamic.TollRateGroupDao;
-import gov.wa.wsdot.android.wsdot.database.tollrates.dynamic.TollRateSignDao;
-import gov.wa.wsdot.android.wsdot.database.tollrates.dynamic.TollRateSignEntity;
-import gov.wa.wsdot.android.wsdot.database.tollrates.dynamic.TollTripDao;
+import gov.wa.wsdot.android.wsdot.database.tollrates.constant.TollRateTable;
+import gov.wa.wsdot.android.wsdot.database.tollrates.constant.TollRateTableDao;
+import gov.wa.wsdot.android.wsdot.database.tollrates.constant.tolltable.TollRateTableDataDao;
+import gov.wa.wsdot.android.wsdot.database.tollrates.constant.tolltable.TollRateTableDataEntity;
+import gov.wa.wsdot.android.wsdot.database.tollrates.constant.tolltable.tollrows.TollRowDao;
+import gov.wa.wsdot.android.wsdot.database.tollrates.constant.tolltable.tollrows.TollRowEntity;
 import gov.wa.wsdot.android.wsdot.util.APIEndPoints;
 import gov.wa.wsdot.android.wsdot.util.network.ResourceStatus;
 import gov.wa.wsdot.android.wsdot.util.threading.AppExecutors;
@@ -37,43 +35,28 @@ public class TollRatesRepository extends NetworkResourceSyncRepository {
 
     private final static String TAG = TollRatesRepository.class.getSimpleName();
 
-    private final TollTripDao tollTripDao;
-    private final TollRateSignDao tollRateSignDao;
-    private final TollRateGroupDao tollRateGroupDao;
+    private final TollRateTableDao tollRateTableDao;
+    private final TollRateTableDataDao tollRateTableDataDao;
+    private final TollRowDao tollRowDao;
 
     @Inject
-    public TollRatesRepository(TollTripDao tollTripDao,
-                               TollRateSignDao tollRateSignDao,
-                               TollRateGroupDao tollRateGroupDao,
+    public TollRatesRepository(TollRateTableDao tollRateTableDao,
+                               TollRateTableDataDao tollRateTableDataDao,
+                               TollRowDao tollRowDao,
                                AppExecutors appExecutors, CacheRepository cacheRepository) {
         super(appExecutors, cacheRepository, DateUtils.MINUTE_IN_MILLIS, "toll_trip");
 
-        this.tollTripDao = tollTripDao;
-        this.tollRateSignDao = tollRateSignDao;
-        this.tollRateGroupDao = tollRateGroupDao;
+        this.tollRateTableDao = tollRateTableDao;
+        this.tollRateTableDataDao = tollRateTableDataDao;
+        this.tollRowDao = tollRowDao;
 
     }
 
-    public LiveData<List<TollRateGroup>> loadI405TollRateGroups(MutableLiveData<ResourceStatus> status) {
+    public LiveData<TollRateTable> loadTollRatesFor(int route, MutableLiveData<ResourceStatus> status) {
         super.refreshData(status, false);
-        return tollRateGroupDao.loadI405TollRateGroups();
+        return tollRateTableDao.loadTollRatesFor(route);
     }
 
-    public LiveData<List<TollRateGroup>> loadSR167TollRateGroups(MutableLiveData<ResourceStatus> status) {
-        super.refreshData(status, false);
-        return tollRateGroupDao.loadSR167TollRateGroups();
-    }
-
-    public LiveData<List<TollRateGroup>> loadFavoriteTolls(MutableLiveData<ResourceStatus> status) {
-        super.refreshData(status, false);
-        return tollRateGroupDao.loadFavoriteTollRateGroups();
-    }
-
-    public void setIsStarred(String location_name, Integer isStarred) {
-        getExecutor().diskIO().execute(() -> {
-            tollRateSignDao.updateIsStarred(location_name, isStarred);
-        });
-    }
 
     /**
      * Builds Toll sign and trip items from API.
@@ -82,18 +65,16 @@ public class TollRatesRepository extends NetworkResourceSyncRepository {
      * sign items and trip items are group together upon retrieval from Room by
      * the sign's location and the trip's start location
      *
-    */
+     */
     @Override
     void fetchData(MutableLiveData<ResourceStatus> status) throws Exception {
 
-        DateFormat dateFormat = new SimpleDateFormat("MMMM d, yyyy h:mm a");
-
-        List<TollRateSignEntity> mTollRateItems = new ArrayList<>();
-        List<TollRateGroupDao.TollTripEntity> mTollTrips = new ArrayList<>();
+        List<TollRateTableDataEntity> mTollTables = new ArrayList<>();
+        List<TollRowEntity> mTollRows = new ArrayList<>();
 
         URL url;
 
-        url = new URL(APIEndPoints.I405_TOLL_RATES + "?accesscode=" + APIEndPoints.WSDOT_API_KEY);
+        url = new URL(APIEndPoints.TOLL_RATES);
         URLConnection urlConn = url.openConnection();
         BufferedReader in = new BufferedReader(new InputStreamReader(urlConn.getInputStream()));
         String jsonFile = "";
@@ -107,219 +88,50 @@ public class TollRatesRepository extends NetworkResourceSyncRepository {
 
         int numItems = items.length();
 
-        for (int j=0; j < numItems; j++) {
+        for (int i = 0; i < numItems; i++) {
 
-            JSONObject item = items.getJSONObject(j);
+            JSONObject item = items.getJSONObject(i);
 
-            TollRateGroupDao.TollTripEntity trip = new TollRateGroupDao.TollTripEntity();
+            TollRateTableDataEntity table = new TollRateTableDataEntity();
 
-            if (!shouldSkipTrip(item)) {
+            table.setRoute(item.getInt("route"));
+            table.setMessage(item.getString("message"));
+            table.setNumCol(item.getInt("numCol"));
 
-                trip.setTripName(item.getString("TripName"));
-                trip.setEndLocationName(item.getString("EndLocationName"));
-                trip.setSignId(item.getString("StartLocationName").concat(item.getString("TravelDirection")));
-                trip.setTollRate(item.getDouble("CurrentToll"));
-                trip.setMessage(item.getString("CurrentMessage"));
-                trip.setEndMilepost(item.getInt("EndMilepost"));
-                trip.setEndLatitude(item.getDouble("EndLatitude"));
-                trip.setEndLongitude(item.getDouble("EndLongitude"));
+            JSONArray rows = new JSONArray(item.getString("tollTable"));
+            int numRows = rows.length();
 
-                try {
-                    trip.setUpdated(dateFormat.format(new Date(System.currentTimeMillis())));
-                } catch (Exception e) {
-                    trip.setUpdated("unavailable");
-                    Log.e(TAG, "Error parsing date", e);
-                }
+            for (int j = 0; j < numRows; j++) {
 
-                mTollTrips.add(trip);
+                TollRowEntity row = new TollRowEntity();
+                JSONObject rowJson = rows.getJSONObject(j);
 
-                TollRateSignEntity sign = tollRateSignDao.getTollRateSign(trip.getSignId());
+                row.setHeader(rowJson.getBoolean("header"));
+                row.setWeekday(rowJson.getBoolean("true"));
 
-                if (sign == null) {
-                    sign = new TollRateSignEntity();
-                }
-                sign.setId(item.getString("StartLocationName").concat(item.getString("TravelDirection")));
+                row.setStartTime(rowJson.getString("start_time"));
+                row.setEndTime(rowJson.getString("end_time"));
 
-                String locationName = item.getString("StartLocationName");
+                row.setRowValues(Converters.fromJsonString(rowJson.getString("rows")));
 
-                if (item.getInt("StateRoute") == 405){
-                    locationName = filter405LocationName(item.getString("StartLocationName"), item.getString("TravelDirection"));
-                } else if (item.getInt("StateRoute") == 167){
-                    locationName = filter167LocationName(item.getString("StartLocationName"), item.getString("TravelDirection"));
-                }
+                mTollRows.add(row);
 
-                sign.setLocationName(locationName);
-                sign.setMilepost(item.getInt("StartMilepost"));
-                sign.setStateRoute(item.getInt("StateRoute"));
-                sign.setTravelDirection(item.getString("TravelDirection"));
-                sign.setStartLatitude(item.getDouble("StartLatitude"));
-                sign.setStartLongitude(item.getDouble("StartLongitude"));
-
-                mTollRateItems.add(sign);
             }
+            mTollTables.add(table);
         }
 
+        TollRateTableDataEntity[] tollTableArray = new TollRateTableDataEntity[mTollTables.size()];
+        tollTableArray = mTollTables.toArray(tollTableArray);
 
-        TollRateSignEntity[] tollRateSignsArray = new TollRateSignEntity[mTollRateItems.size()];
-        tollRateSignsArray = mTollRateItems.toArray(tollRateSignsArray);
+        tollRateTableDataDao.deleteAndInsertTransaction(tollTableArray);
 
-        tollRateSignDao.deleteAndInsertTransaction(tollRateSignsArray);
+        TollRowEntity[] tollRowArray = new TollRowEntity[mTollRows.size()];
+        tollRowArray = mTollRows.toArray(tollRowArray);
 
-        TollRateGroupDao.TollTripEntity[] tollTripsArray = new TollRateGroupDao.TollTripEntity[mTollTrips.size()];
-        tollTripsArray = mTollTrips.toArray(tollTripsArray);
+        tollRowDao.deleteAndInsertTransaction(tollRowArray);
 
-        tollTripDao.deleteAndInsertTransaction(tollTripsArray);
-
-        CacheEntity tollCache = new CacheEntity("toll_trip", System.currentTimeMillis());
+        CacheEntity tollCache = new CacheEntity("toll_table", System.currentTimeMillis());
         getCacheRepository().setCacheTime(tollCache);
 
-    }
-
-    private String filter405LocationName(String locationName, String direction){
-
-        // Southbound name changes suggested by Tolling
-        if (direction.equals("S")) {
-
-            if (locationName.equals("231st SE")) {
-                locationName = "SR 527";
-            }
-
-            if (locationName.equals("NE 53rd")) {
-                locationName = "NE 70th Place";
-            }
-
-        }
-
-        // Northbound name changes suggested by Tolling
-        if (direction.equals("N")) {
-
-            if (locationName.equals("NE 97th")) {
-                locationName = "NE 85th St";
-            }
-
-            if (locationName.equals("231st SE")) {
-                locationName = "SR 522";
-            }
-
-            if (locationName.equals("216th SE")) {
-                locationName = "SR 527";
-            }
-        }
-
-        if (locationName.equals("SR 524") || locationName.equals("NE 4th")){
-            locationName = (direction.equals("N") ? "Bellevue" : "Lynnwood").concat(" - Start of toll lanes");
-        } else {
-            locationName = "Lane entrance near ".concat(locationName);
-        }
-
-        return locationName;
-    }
-
-
-    private String filter167LocationName(String locationName, String direction){
-
-        // Southbound name changes suggested by Tolling
-        if (direction.equals("S")) {
-
-            if (locationName.equals("4th Ave N")) {
-                locationName = "SR 516";
-            }
-
-            if (locationName.equals("S 192nd St")) {
-                locationName = "S 180th St";
-            }
-
-            if (locationName.equals("S 23rd St")) {
-                locationName = "I-405 (Renton)";
-            }
-        }
-
-        // Northbound name changes suggested by Tolling
-        if (direction.equals("N")) {
-
-            if (locationName.equals("15th St SW")) {
-                locationName = "SR 18 (Auburn)";
-            }
-
-            if (locationName.equals("7th St NW")) {
-                locationName = "15th St SW";
-            }
-
-            if (locationName.equals("30th St NW")) {
-                locationName = "S 277th St";
-            }
-
-            if (locationName.equals("S 265th St")) {
-                locationName = "SR 516";
-            }
-        }
-
-        locationName = "Lane entrance near ".concat(locationName);
-
-        return locationName;
-    }
-
-    private boolean shouldSkipTrip(JSONObject tripJson) throws JSONException {
-
-        /*
-         * Removal of these routes since their displays are already shown
-         * by other signs from the API.
-         */
-        if (tripJson.getString("StartLocationName").equals("NE 6th")
-            && tripJson.getString("TravelDirection").equals("N")) {
-            return true;
-        }
-
-        if (tripJson.getString("StartLocationName").equals("216th ST SE")
-                && tripJson.getString("TravelDirection").equals("S")) {
-            return true;
-        }
-
-        if (tripJson.getString("StartLocationName").equals("NE 145th")
-                && tripJson.getString("TravelDirection").equals("S")) {
-            return true;
-        }
-
-        /*
-         * Removal suggested by tolling division since it's very similar to another location
-         * and difficult to come up with a label people will recognize.
-         */
-        if (tripJson.getString("StartLocationName").equals("NE 108th")
-                && tripJson.getString("TravelDirection").equals("S")) {
-            return true;
-        }
-
-        // SR 167 trips to remove
-        if (tripJson.getString("StartLocationName").equals("James St")
-                && tripJson.getString("TravelDirection").equals("N")) {
-            return true;
-        }
-
-        if (tripJson.getString("StartLocationName").equals("S 204th St")
-                && tripJson.getString("TravelDirection").equals("N")) {
-            return true;
-        }
-
-        if (tripJson.getString("StartLocationName").equals("1st Ave S")
-                && tripJson.getString("TravelDirection").equals("S")) {
-            return true;
-        }
-
-        if (tripJson.getString("StartLocationName").equals("12th St NW")
-                && tripJson.getString("TravelDirection").equals("S")) {
-            return true;
-        }
-
-        if (tripJson.getString("StartLocationName").equals("37th St NW")
-                && tripJson.getString("TravelDirection").equals("S")) {
-            return true;
-        }
-
-        if (tripJson.getString("StartLocationName").equals("Green River")
-                && tripJson.getString("TravelDirection").equals("S")) {
-            return true;
-        }
-
-        return false;
     }
 }
